Initial commit with Elixir service credits project
This repo is intended to contain a mix of projects that demonstrate some
of the work I've done in the past. Some projects will instead just be
snippets while others might be closer to a full application. I will try
to keep the content within this portfolio limited to only things of
which I had full or majority ownership in, i.e. I did most or all of the
work to create it.